+++ spec/guard/runner_spec.rb
@@ -14,17 +14,21 @@ describe Guard::Runner do
     @foo_guard      = guard.add_plugin(:foo, { :group => :backend })
     @bar_guard      = guard.add_plugin(:bar, { :group => :frontend })
     @baz_guard      = guard.add_plugin(:baz, { :group => :frontend })
+
+    @foo_guard.stub(:my_task)
+    @bar_guard.stub(:my_task)
+    @baz_guard.stub(:my_task)
+
+    @foo_guard.stub(:my_hard_task)
+    @bar_guard.stub(:my_hard_task)
   end
 
   describe '#run' do
-    let(:scopes) { { :group => @backend_group } }
-
     it 'executes a supervised task on all registered plugins implementing that task' do
-      [@foo_guard, @bar_guard].each do |g|
-        g.stub(:my_task)
-        subject.should_receive(:run_supervised_task).with(g, :my_task)
+      [@foo_guard, @bar_guard].each do |plugin|
+        subject.should_receive(:run_supervised_task).with(plugin, :my_hard_task)
       end
-      subject.run(:my_task)
+      subject.run(:my_hard_task)
     end
 
     it 'marks an action as unit of work' do
@@ -42,31 +46,114 @@ describe Guard::Runner do
       end
     end
 
-    context 'within the scope of a specified local guard' do
-      let(:scopes) { { :plugins => [@bar_guard] } }
+    context 'with a plugin as scope' do
+      context 'passed as a symbol' do
+        let(:scope) { { :plugin => :bar } }
 
-      it 'executes the supervised task on the specified guard only' do
-        @bar_guard.stub(:my_task)
-        subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+        it 'executes the supervised task on the specified plugin only' do
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
 
-        subject.should_not_receive(:run_supervised_task).with(@foo_guard, :my_task)
-        subject.should_not_receive(:run_supervised_task).with(@baz_guard, :my_task)
+          subject.should_not_receive(:run_supervised_task).with(@foo_guard, :my_task)
+          subject.should_not_receive(:run_supervised_task).with(@baz_guard, :my_task)
 
-        subject.run(:my_task, scopes)
+          subject.run(:my_task, scope)
+        end
+      end
+
+      context 'passed as a Guard::Plugin object' do
+        let(:scope) { { :plugin => @bar_guard } }
+
+        it 'executes the supervised task on the specified plugin only' do
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+
+          subject.should_not_receive(:run_supervised_task).with(@foo_guard, :my_task)
+          subject.should_not_receive(:run_supervised_task).with(@baz_guard, :my_task)
+
+          subject.run(:my_task, scope)
+        end
       end
     end
 
-    context 'within the scope of a specified local group' do
-      let(:scopes) { { :groups => [@backend_group] } }
+    context 'with an array of plugins as scope' do
+      context 'passed as a symbol' do
+        let(:scope) { { :plugins => [:foo, :bar] } }
 
-      it 'executes the task on each guard in the specified group only' do
-        @foo_guard.stub(:my_task)
-        subject.should_receive(:run_supervised_task).with(@foo_guard, :my_task)
+        it 'executes the supervised task on the specified plugins only' do
+          @bar_guard.stub(:my_task)
+          subject.should_receive(:run_supervised_task).with(@foo_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
 
-        subject.should_not_receive(:run_supervised_task).with(@bar_guard, :my_task)
-        subject.should_not_receive(:run_supervised_task).with(@baz_guard, :my_task)
+          subject.should_not_receive(:run_supervised_task).with(@baz_guard, :my_task)
 
-        subject.run(:my_task, scopes)
+          subject.run(:my_task, scope)
+        end
+      end
+
+      context 'passed as a Guard::Plugin objects' do
+        let(:scope) { { :plugins => [@foo_guard, @bar_guard] } }
+
+        it 'executes the supervised task on the specified plugins only' do
+          subject.should_receive(:run_supervised_task).with(@foo_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+
+          subject.should_not_receive(:run_supervised_task).with(@baz_guard, :my_task)
+
+          subject.run(:my_task, scope)
+        end
+      end
+    end
+
+    context 'with a group as scope' do
+      context 'passed as a symbol' do
+        let(:scope) { { :group => :frontend } }
+
+        it 'executes the supervised task on the specified plugin only' do
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@baz_guard, :my_task)
+
+          subject.should_not_receive(:run_supervised_task).with(@foo_guard, :my_task)
+
+          subject.run(:my_task, scope)
+        end
+      end
+
+      context 'passed as a Guard::Group object' do
+        let(:scope) { { :group => @frontend_group } }
+
+        it 'executes the supervised task on the specified plugin only' do
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@baz_guard, :my_task)
+
+          subject.should_not_receive(:run_supervised_task).with(@foo_guard, :my_task)
+
+          subject.run(:my_task, scope)
+        end
+      end
+    end
+
+    context 'with an array of plugins as scope' do
+      context 'passed as a symbol' do
+        let(:scope) { { :groups => [:frontend, :backend] } }
+
+        it 'executes the supervised task on the specified plugins only' do
+          subject.should_receive(:run_supervised_task).with(@foo_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@baz_guard, :my_task)
+
+          subject.run(:my_task, scope)
+        end
+      end
+
+      context 'passed as a Guard::Group objects' do
+        let(:scope) { { :groups => [@frontend_group, @backend_group] } }
+
+        it 'executes the supervised task on the specified plugins only' do
+          subject.should_receive(:run_supervised_task).with(@foo_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@bar_guard, :my_task)
+          subject.should_receive(:run_supervised_task).with(@baz_guard, :my_task)
+
+          subject.run(:my_task, scope)
+        end
       end
     end
   end